[INTERNAL] Use local replacement for sub-modules
As part of https://github.com/banzaicloud/koperator/pull/929, the local
replacement for sub-modules was removed in favor of using valid tags.
Adobe Koperator fork has some internal changes in the api sub-module,
so we need to use the local version of the sub-module instead of
the upstream version, so we are  forced to revert the changes from #929.